LayoutTests/inspector/animation/targetChanged-expected.txt

Tests for the Animation.targetChanged event.


== Running test suite: Animation.targetChanged
-- Running test case: Animation.targetChanged.NewTarget
PASS: Animation effect should have a target element.
PASS: Animation effect should not have a target pseudo-element.
Changing target...

PASS: Animation effect should have a target element.
PASS: Animation effect should not have a target pseudo-element.
PASS: Animation effect target should have changed.
PASS: Animation effect target element should have changed.

-- Running test case: Animation.targetChanged.NullTarget
PASS: Animation effect should have a target element.
PASS: Animation effect should not have a target pseudo-element.
Changing target...

PASS: Animation effect should not have a target element.
PASS: Animation effect target should have changed.

-- Running test case: Animation.targetChanged.NewTargetWithPseudoId
PASS: Animation effect should not have a target.
Changing target...

PASS: Animation effect should have a target element.
PASS: Animation effect should not have a target pseudo-element.
Changing target again...

PASS: Animation effect should have a target element.
PASS: Animation effect should have a target pseudo-element.
PASS: Animation effect target should have changed.
PASS: Animation effect target element should not have changed.
